requirements.txt
Jinja2==2.10.3
parsec==3.5.1
PyGithub==1.44.1
python-dateutil==2.6.1
PyYAML==5.1
requests==2.22.0
